Added exchange method

+++ src/teachmeskills/vlad/hw14/currencytype/ExchangeRate.java
@@ -1,10 +1,12 @@
 package teachmeskills.vlad.hw14.currencytype;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
     private Currency currencyType;
-    private double course;
+    private BigDecimal course;
 
-    ExchangeRate(Currency currencyType, double course) {
+    ExchangeRate(Currency currencyType, BigDecimal course) {
         this.currencyType = currencyType;
         this.course = course;
     }
@@ -13,7 +15,7 @@ public class ExchangeRate {
         return currencyType;
     }
 
-    public double getCourse() {
+    public BigDecimal getCourse() {
         return course;
     }
 }
